server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
openai==1.3.7
httpx==0.25.2
python-multipart==0.0.6
aiofiles==23.2.1
